requirements.txt
pandas
numpy
beautifulsoup4
requests
pybaseball
unidecode
gspread
gspread-dataframe
google-auth
MLB-StatsAPI
jinja2